Split out trace_counter mod
